+++ src/app/src/components/ClaimedFacilitiesDetails.jsx
@@ -13,7 +13,6 @@ import memoize from 'lodash/memoize';
 import find from 'lodash/find';
 import stubFalse from 'lodash/stubFalse';
 import isEmpty from 'lodash/isEmpty';
-import isEqual from 'lodash/isEqual';
 import get from 'lodash/get';
 import map from 'lodash/map';
 import filter from 'lodash/filter';
@@ -49,8 +48,7 @@ import {
     updateClaimedFacilityDescription,
     updateClaimedFacilityMinimumOrder,
     updateClaimedFacilityAverageLeadTime,
-    updateClaimedFacilityFacilityType,
-    updateClaimedFacilityOtherFacilityType,
+    updateClaimedFacilityFacilityTypes,
     updateClaimedFacilityContactPersonName,
     updateClaimedFacilityContactEmail,
     updateClaimedFacilityPointOfContactVisibility,
@@ -298,7 +296,6 @@ function ClaimedFacilitiesDetails({
     updateParentCompany,
     contributorOptions,
     updateFacilityType,
-    updateFacilityOtherType,
 }) {
     /* eslint-disable react-hooks/exhaustive-deps */
     // disabled because we want to use this as just
@@ -428,24 +425,15 @@ function ClaimedFacilitiesDetails({
                 </ShowOnly>
                 <InputSection
                     label="Facility Type"
-                    value={get(data, 'facility_type', null)}
+                    value={get(data, 'facility_type', [])}
                     onChange={updateFacilityType}
                     disabled={updating}
                     isSelect
+                    isMultiSelect
                     selectOptions={mapDjangoChoiceTuplesToSelectOptions(
                         data.facility_types,
                     )}
                 />
-                <ShowOnly
-                    when={isEqual(get(data, 'facility_type', null), 'Other')}
-                >
-                    <InputSection
-                        label="Other Facility Type"
-                        value={get(data, 'other_facility_type', null)}
-                        onChange={updateFacilityOtherType}
-                        disabled={updating}
-                    />
-                </ShowOnly>
                 <InputSection
                     label="Minimum order quantity"
                     value={data.facility_minimum_order_quantity}
@@ -522,13 +510,12 @@ function ClaimedFacilitiesDetails({
                     )}
                 />
                 <InputSection
-                    label="Production Types"
+                    label="Processing Types"
                     value={get(data, 'facility_production_types', [])}
                     onChange={updateFacilityProductionTypes}
                     disabled={updating}
                     isSelect
                     isMultiSelect
-                    isCreatable
                     selectOptions={mapDjangoChoiceTuplesToSelectOptions(
                         data.production_type_choices,
                     )}
@@ -695,7 +682,6 @@ ClaimedFacilitiesDetails.propTypes = {
     updateOfficeVisibility: func.isRequired,
     contributorOptions: contributorOptionsPropType,
     updateFacilityType: func.isRequired,
-    updateFacilityOtherType: func.isRequired,
 };
 
 function mapStateToProps({
@@ -775,10 +761,8 @@ function mapDispatchToProps(
         updateFacilityDescription: makeDispatchValueFn(
             updateClaimedFacilityDescription,
         ),
-        updateFacilityType: ({ value }) =>
-            dispatch(updateClaimedFacilityFacilityType(value)),
-        updateFacilityOtherType: makeDispatchValueFn(
-            updateClaimedFacilityOtherFacilityType,
+        updateFacilityType: makeDispatchMultiSelectFn(
+            updateClaimedFacilityFacilityTypes,
         ),
         updateFacilityMinimumOrder: makeDispatchValueFn(
             updateClaimedFacilityMinimumOrder,